# docs/openapi.yaml
openapi: 3.1.0
info:
  title: stock-market-simulator API
  version: 0.1.0
  summary: REST API simulating a simplified stock exchange with a single bank as the sole liquidity source.
  description: |
    Eight endpoints over a single Postgres backing store. Buy and sell are
    atomic with audit-log writes (see ADR 0005). The HA setup runs N app
    replicas behind Caddy with DNS-based discovery (see ADR 0001 / 0002).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local compose stack (Caddy in front of N app replicas)
security: []
paths:
...
  /wallets/{wallet_id}:
    get:
      summary: Read a wallet and its holdings
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Wallet state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wallets/{wallet_id}/stocks/{stock_name}:
    get:
      summary: Read a single stock holding for a wallet
      description: |
        Returns the quantity as a bare JSON integer (not an object). Returns 0
        when the wallet exists but has no holding for that stock.
      operationId: getWalletStockQuantity
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/StockName'
      responses:
        '200':
          description: Holding quantity as a bare JSON integer
          content:
            application/json:
              schema:
                type: integer
                format: int64
                minimum: 0
                example: 99
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Buy or sell one unit of a stock for a wallet
      description: |
        One unit moves between the bank and the wallet, with the audit row
        written in the same database transaction. The wallet is created on
        the first buy or sell and persists even when the operation itself
        fails.
      operationId: buyOrSell
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/StockName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
      responses:
        '200':
          description: Operation completed (no body)
        '400':
          description: Invalid operation, insufficient bank stock, or insufficient wallet stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Stock is not present in the bank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    WalletId:
      name: wallet_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        example: w1
    StockName:
      name: stock_name
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        example: AAPL
  schemas:
    Stock:
      type: object
      required: [name, quantity]
      properties:
        name:
          type: string
          minLength: 1
          example: AAPL
        quantity:
          type: integer
          format: int64
          minimum: 0
          example: 100
    Wallet:
      type: object
      required: [id, stocks]
      properties:
        id:
          type: string
          example: w1
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/Stock'
...
    OperationRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [buy, sell]
          example: buy
    Error:
      type: object
      description: Standard error envelope. `code` is the stable contract; `error` is the human-readable message and may change.
      required: [error, code]
      properties:
        error:
          type: string
          example: wallet not found
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - INVALID_OPERATION
            - INVALID_QUANTITY
            - EMPTY_STOCK_NAME
            - DUPLICATE_STOCK_NAME
            - INSUFFICIENT_BANK_STOCK
            - INSUFFICIENT_WALLET_STOCK
            - WALLET_NOT_FOUND
            - STOCK_NOT_FOUND
            - INTERNAL_ERROR
          example: WALLET_NOT_FOUND